Prefiltering on points. Switching over to int64 timestamps
There's a lot in this commit! Dotmaster has had its time range methods
expanded with Intersections for time ranges and disjoint ranges. This is
used to prefilter a data query so that instead of querying for all the
data a user has asked for and then iterating through those points to
filter out those from times that are inaccessible by that user, we
simply rewrite the user's query so its in the bounds of what they're
allowed to see.

The danger of doing this is that we lose out an some natural parallelism
if we dispatch a bunch of UUIDs to BtrDB and ask for the same range for
all of them; this is potentially a bit faster (and we could introduce a
special case for that), but this is already netting us a 30-50% increase
in query speed, which is great

+++ archiver/transforms.go
@@ -37,7 +37,7 @@ func POsFromTimeseriesGroup(nonce uint32, tsGroups []common.Timeseries, statsGro
 			UUID:       group.UUID.String(),
 			Path:       group.SrcURI,
 			Generation: group.Generation,
-			Times:      []uint64{},
+			Times:      []int64{},
 			Values:     []float64{},
 		}
 		for _, rdg := range group.Records {
@@ -57,7 +57,7 @@ func POsFromTimeseriesGroup(nonce uint32, tsGroups []common.Timeseries, statsGro
 			Max:        []float64{},
 		}
 		for _, rdg := range group.Records {
-			ts.Times = append(ts.Times, common.TimeAsUnit(rdg.Time, rdg.Unit))
+			ts.Times = append(ts.Times, uint64(common.TimeAsUnit(rdg.Time, rdg.Unit)))
 			ts.Count = append(ts.Count, rdg.Count)
 			ts.Min = append(ts.Min, rdg.Min)
 			ts.Mean = append(ts.Mean, rdg.Mean)
